zbd: add explicit open support for zns-zrwa drives.
ZRWA spec allows the zones to have a random write area (RWA), which
starts at the WP and ends after a certain boundary. This area
keeps moving down along with the WP like a moving window till
the end of zone (capacity). Within this RWA, overwrites are allowed
(on a seq write only zone). This patch series allows fio to open
the zones in zrwa mode and issue commit zone command and overwrites
to the zones.
For zones to be opened with ZRWA attached, it requires
the zone to be opened explicitly and send a ZRWAA flag to the
drive.

Signed-off-by: Aravind Ramesh <Aravind.Ramesh@wdc.com>