zbd: Add over write support to sequential zones.
Allow fio to issue overwrites as a percentage of zone size.
zrwa_overwrite_percent option indicates the amount of
overwrites in percentage of zone size. FIO issues writes
to previously written lbas. This may need fine tuning
considering the ZRWA boundary is changing continuously
and the issued overwrite may go to a commited part of the
zone leading to write error, recommended to not use along
with zrwa explicit commit option.

Signed-off-by: Aravind Ramesh <Aravind.Ramesh@wdc.com>