zbd: read with verify failure
During a read with verify enabled workload, fio was calling
write_replay function and it was not able to open zone and
zone-capacity was not getting handled.
Now, zbd_adjust_block() will check the zone-capacity checks
and then pass the correct zone, for write_replay cases.
Reported by Jothan Tu and team.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>