zbd: dynamic qd implementation
Maintain a dynamic qd for zones when zrwa is enabled.
Fio will not submit an IO to the next lba, if any already
submitted IOs are still not complete and if submitting
the next IO can cause any previously submitted IOs to fail.
This is done by taking ZRWA SIZE into consideration.
For ex: With 256K bs, and qd=4, fio will initially submit
4 IOs (to index 0,1,2,3). With this change, fio will not
submit write IO to index 4, until IO to index 0 is completed.
To clarify further, if IO to index 0 completes first, then fio
will submit IO to index 4, if not, it will wait even if IOs to
index 1,2 and 3 completes and once IO to index 0 completes,
then fio will submit IO to index 4. If IO to index 4 is
submitted before IO to index 0 is complete, then there is
chance where 4 completes first moving the device WP to index 1
and causing the IO to index 0 to fail.

While sending finish zone, fio will have to wait until all
queued IOs to that zone are completed and then finish the zone.

With this patch, maintaining the qd at an individual zone level
is possible.
For ex: For a zone with ZRWA area of 1MB, with bs=256K we can have
a qd of 4 for that zone. So with rw=randwrite, max_open_zones=12
and bs=256K we can set iodepth=48 (12 * 4), fio maintains dynamic qd to
zones ensuring writes don't go behind zone WP, so it will not queue
more than 4 IOs per zone (for ZRWA=1MB and bs=256K), which will
result in all 12 zones getting a maximum of 4 qd.
Without this patch fio can queue more IOs to a zone such that
it will cross the ZRWA region causing write errors.

This patch assumes ZRWA=1MB.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>
Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>